requirements.txt
altgraph==0.17.4
beautifulsoup4==4.13.4
lxml==5.4.0
macholib==1.16.3
markdownify==1.1.0
packaging==25.0
pyinstaller==6.14.1
pyinstaller-hooks-contrib==2025.5
PyMuPDF==1.26.1
PySide6==6.9.1
PySide6_Addons==6.9.1
PySide6_Essentials==6.9.1
python-docx==1.2.0
setuptools==80.9.0
shiboken6==6.9.1
six==1.17.0
soupsieve==2.7
typing_extensions==4.14.0
